adicionando novos métodos para os cadastros de livro

+++ 03_RestWithASPNet_CalculatorControllersAndModels/RestWithASPNet/RestWithASPNet/Business/IPersonBusiness.cs
@@ -6,12 +6,12 @@ namespace RestWithASPNet.Business
     {
         Person Create(Person person);
 
-        Person FindById(long id);
+        Person FindById(int id);
 
         List<Person> FindAll();
 
         Person Update(Person person);
 
-        void Delete(long id);
+        void Delete(int id);
     }
 }
