+++ 03_RestWithASPNet_CalculatorControllersAndModels/RestWithASPNet/RestWithASPNet/Repository/IPersonRepository.cs
@@ -6,14 +6,14 @@ namespace RestWithASPNet.Repository
     {
         Person Create(Person person);
 
-        Person FindById(long id);
+        Person FindById(int id);
 
         List<Person> FindAll();
 
         Person Update(Person person);
 
-        void Delete(long id);
+        void Delete(int id);
 
-        bool Exists(long id);
+        bool Exists(int id);
     }
 }
